test(ui): run announcement keyboard and reduced-motion scenarios in required E2E
