+++ scripts/browser/core.mjs
@@ -1,4 +1,5 @@
 import assert from "node:assert/strict";
+import { inspectAnnouncements } from "./announcements.mjs";
 import { inspectReceipts } from "./receipts.mjs";
 import { inspectEpea } from "./epea.mjs";
 import { inspectCulture, inspectCultureJourney, inspectInvalidCultureYears } from "./culture.mjs";
@@ -698,6 +699,14 @@ const completed = [];
 
 try {
   browser = await launchBrowser();
+
+  for (const width of [390, 1440]) {
+    const label = `Annunci accessibili ${width}px`;
+    await runScenario(browser, {
+      label, pathname: "/", width, validate: inspectAnnouncements,
+    });
+    completed.push(label);
+  }
 
   for (const width of [320, 375, 390, 768, 1024, 1280, 1600]) {
     const label = `Palette istituzionale ${width}px`;
